refactor: refactor existing commands to work with new implementation
